requirements.txt
Django==3.1.7
psycopg2-binary==2.8.6
pylint==2.7.2
djangorestframework==3.12.2
django-rest-swagger==2.2.0
pyjwt==2.0.1
